+++ source/F0.Common/F0.Common/Mathematics/Parity.cs
@@ -4,12 +4,50 @@
 	{
 		public static bool IsEven(int integer)
 		{
-			return integer % 2 == 0;
+			int remainder = integer % 2;
+			return remainder == 0;
+		}
+
+		public static bool IsEven(uint integer)
+		{
+			uint remainder = integer % 2u;
+			return remainder == 0u;
+		}
+
+		public static bool IsEven(long integer)
+		{
+			long remainder = integer % 2L;
+			return remainder == 0L;
+		}
+
+		public static bool IsEven(ulong integer)
+		{
+			ulong remainder = integer % 2ul;
+			return remainder == 0ul;
 		}
 
 		public static bool IsOdd(int integer)
 		{
-			return integer % 2 != 0;
+			int remainder = integer % 2;
+			return remainder != 0;
+		}
+
+		public static bool IsOdd(uint integer)
+		{
+			uint remainder = integer % 2u;
+			return remainder == 1u;
+		}
+
+		public static bool IsOdd(long integer)
+		{
+			long remainder = integer % 2L;
+			return remainder != 0L;
+		}
+
+		public static bool IsOdd(ulong integer)
+		{
+			ulong remainder = integer % 2ul;
+			return remainder == 1ul;
 		}
 	}
 }