+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 CHANGELOG
 
 ## vNext
+- Added `Mathematics.Parity` checks for `uint`, `long` and `ulong`.
 
 ## v0.3.0 (2019-04-30)
 - Changed target framework from `.NET Standard 1.0` to `.NET Standard 2.0`.
@@ -11,4 +12,4 @@ CHANGELOG
 - Added `Primitives.NullProgress<T>`, a Null Object implementation of `System.IProgress<T>` as Singleton.
 
 ## v0.1.0 (2018-09-18)
-- Added `Mathematics.Parity` checking whether an integer is even or odd.
+- Added `Mathematics.Parity`, checking whether an `int` is even or odd.